+++ libguile/filesys.c
@@ -751,7 +751,7 @@ scm_dir_print (SCM exp, SCM port, scm_print_state *pstate)
   if (SCM_CLOSEDP (exp))
     scm_puts ("closed: ", port);
   scm_puts ("directory stream ", port);
-  scm_intprint (SCM_CDR (exp), 16, port);
+  scm_intprint ((int)SCM_CDR (exp), 16, port);
   scm_putc ('>', port);
   return 1;
 }
